requirements.txt
# Core dependencies for Google Slides Crawler
selenium>=4.15.0
google-cloud-vision>=3.4.0
google-generativeai>=0.3.0

# Web automation and browser control
webdriver-manager>=4.0.0

# Image processing (optional, for image preprocessing)
Pillow>=10.0.0

# HTTP requests and API handling
requests>=2.31.0

# Environment and configuration management
python-dotenv>=1.0.0

# Logging and monitoring
colorlog>=6.7.0

# Testing dependencies (optional)
pytest>=7.4.0
pytest-selenium>=4.0.0

# Development dependencies (optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation (optional)
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0 